refactor(comp/checks/agentcrashdetect): migrate to V2 component pattern
Split into def/impl/fx/mock packages following the V2 component
convention. Convert fx.Lifecycle/fx.Hook to compdef equivalents
and update callers to use agentcrashdetectfx.Module().
