+++ comp/checks/agentcrashdetect/def/component.go
